Reorganized the num-saving test suite
FlowchartSavingTests was getting a bit big, and I felt that testing other var type encodings would take similar amounts of code. Hence, I made FlowchartSavingTests an abstract class for various suites that each focus on certain var types or groups.

Hence the new FlowchartNumSavingTests suite.
