+++ Assets/Tests/PlayMode/SaveSystem/FlowchartNumVarSavingTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d630f6ea68be60c46a43644ea622d57c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
